refactor wagmi code into own file
