+++ src/features/DecentSwap/DecentSwap.tsx
@@ -3,9 +3,9 @@ import '@decent.xyz/box-ui/index.css';
 import { BoxHooksContextProvider } from '@decent.xyz/box-hooks';
 import { WagmiProvider } from 'wagmi';
 
+import { wagmiConfig } from '../wagmi/config';
 import { DECENT_XYZ_API_KEY } from 'config/env';
 
-import { wagmiConfig } from './config';
 import { SwapComponent } from './SwapComponent';
 
 export function DecentSwap() {
